+++ obv_messenger/app/src/main/java/io/olvid/messenger/databases/tasks/HandleNewMessageNotificationTask.java
@@ -642,7 +642,7 @@ public class HandleNewMessageNotificationTask implements Runnable {
 
                 engine.markMessageForDeletion(obvMessage.getBytesToIdentity(), obvMessage.getIdentifier());
 
-                if (fyleMessageJoinWithStatusesToDownload.size() > 0) {
+                if (!fyleMessageJoinWithStatusesToDownload.isEmpty()) {
                     AvailableSpaceHelper.refreshAvailableSpace(false);
                 }